requirements.txt
SQLAlchemy==2.0.7
alembic==1.10.2
logram==0.0.4
python-dotenv==1.0.0
psycopg2==2.9.5
